Revert "osbuild: ensure loop.Loop() has the required device node"
This reverts commit 158acaac785056132bffa777755d93e490e2e2d6.

With https://github.com/osbuild/bootc-image-builder/pull/238 the
original reason to call mknod goes away so we can just revert
it. osbuild now requires not only the loop device but also uses
`losetup --partscan` quite a lot now so the mknod approach becomes
impractical and the consumers of osbuild in a container should
just setup devtmpfs.