tools/osbuild-depsolve-dnf: repo info on package
When generating package sources and rpm stage metadata for a manifest
from a list of packages, we need to associate repository configuration
options to each package [1].

Add the information to the packages in osbuild-depsolve-dnf before
returning the packages so we don't have to do the association in the
user.  This is needed for reading repository configurations from a
directory so that the caller doesn't need to parse the configurations
itself.

[1] https://github.com/osbuild/images/blob/92497c7b1fcf8a09d136368db031ab65eedd7fa2/pkg/dnfjson/dnfjson.go#L499-L507