workflow,pytest: add new `--error-on-skip` for pytest
To ensure we do not skip tests in GH actions that we want to run
this commit add a new explicit way to allowlist tests that are okay
to skip.

This is done via a small pytest plugin and the GH action will
now use `-k "not test_selinux_ro"` because this selinux tests
requires a real selinux which is not available on GH runners.