sources/curl: pretend to be libdnf in the user agent
Some RPM mirrors (e.g. mirror.cpsc.ucalgary.ca) apparently denylist the
curl user agent. Since we are doing basically the same thing as libdnf,
let's pretend that we are libdnf. However, since the curl source is
generic, we probably don't want to set it for all downloads, just for
the rpm ones. Let's use a simple heurestic based on the file extension
in the URI to determine whether to spoof the agent, or not.