osbuild-depsolve-dnf5: Wait until Fedora 41 to switch to DNF5
Due to dnf5 announcing dnf 5.2.0 with breaking API changes we need to
update the current osbuild-depsolve-dnf5. This would mean carrying 3
different scripts if we support dnf 5.1.x on Fedora 40 so it switches
back to using libdnf for 40, and switches to dnf 5.2.0 for Fedora 41.

NOTE: This needs to wait for libdnf5 version 5.2.0 to be merged into
Fedora 41 (currently rawhide) before it should be merged.

It will also require the Schutzfile osbuild entry for Fedora 40 in the
images repo to be updated so that it is testing the correct depsolver.