tools: use env/marker-file to detect what DNF version to use
The previous approach of detecting what dnf version to use
based on arv0 was not liked very much and it does not fit with
the osbuild.spec that just installs o-d-dnf5 as o-d-dnf.

So this commit switches the approach to use a marker-file instead.
This is just a strawman, it could be an ini file, it could be
something that the spec file prepends to the python script itself
or a ton of other ways, happy to do whatever people like most.

It also has an env to make the tests easier but that is not required,
this can also be simplified/changed and a fake marker file could
be used during the test.