CompletedBuild: separate stdout and stderr
This is the first step towards the `stderr` and `stdout` properties of
the `CompletedBuild` process to be separated. The `output` property is
still a combination of both.

This will make it easier to test output of processes run in the
buildroot, without taking into account the variety of potential error
messages in the combined `output` property.

For now, just make the `CompletedBuild` constructor to accept separate
`stdout` and `stderr` and keep passing the combined `output` as `stdout`
and an empty string as `stderr`.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>