BuildRoot.run(): set separate stdout / stderr in CompletedBuild
Rework the `BuildRoot.run()` method to separately read output from
process stdout and stderr. Both are still logged using the monitor, but
they are set separately in the returned `CompletedBuild` object. This
will make it possible to get the process stdout output, without it being
polluted by the stderr messages.

Simplify `run()` method's part responsible for reading process file
descriptors, so that they are read only within the `while` loop and not
once again after breaking from it. As a result, the
`read_with_timeout()` is removed and its logic is now incorporated by
`run()`.

Add a simple unit test to verify that `stderr` and `stdout` on
`CompletedBuild` object returned by `run()` are separated and that
`output` is a combination of both.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>