stages: run "ksvalidator" optionally when generating kickstarts
There was a bug in osbuild/bootc-image-builder#363 where a stray
newlnine in the ssh key broke the kickstart file.

The isse is that when auto-generating the config.json for an ssh keyfile
it's easy to accidentally include the final `\n` in the ssh key json
string. This will break in very non-obvious ways, i.e. the image will
be generated but the kickstart file is broken because now it looks like:
```
sshkey --username foo "ssh-key
"
```
which make the kickstart invalid but that is only discovered much later
when the image booted (or even later after the installer put the
kickstart into the target system).

This commit runs `ksvalidator` (if available) to detect this kind
of issue early. It requires `pykickstart` in the buildroot but
that seems to be not too bad as we already have `python3-kickstart`
in our buildroots AFAICT (unless I read the manifest-db incorrectly).

So it does require an extra tweak to `images` before it's fully
useful.