buildroot: support OSTree trees as buildroots
In a filesystem layout in OSTree the software is in a deployment
under `/ostree/deploy/<name>/deploy/*/`. If the specified buildroot
is an OSTree tree then let's just use the deployment root as the
rootdir here.

An example where this will help us is in RHCOS, where we'd like to
not maintain a separate buildroot definition since our OSTree we
already build has all the exact versions of the software we want
to use already.