monitor: add new `overall_result` helper to the monitor
One goal of the json-seq mmonitor is that it can be used to get
all relevant information about the build process of an `osbuild`
run. But the overall result was missing, it is printed out-of-band
with `--json` to json but outside of the monitor.

This commit adds a new `overall_result()` method to the
`osbuild.Monitor` that is then used in the json-seq monitor to
output the same json result as `--json` via the new `overall_result`
key in the json-seq output. This will allow osbuild-composer to
use the json-seq monitor to get real-time updates from the worker
executor and at the end read the build result from the same
channel.